src/products-api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.4
elasticsearch==8.12.1
prometheus-client==0.20.0
